linter move code
