+++ packages/generator-common/src/file-writer/copy-file.spec.ts
