+++ packages/generator-common/src/file-writer/index.ts
@@ -1,0 +1,1 @@
+export * from './copy-file';
