+++ packages/generator-common/src/internal.ts
@@ -3,3 +3,4 @@ export * from './compiler';
 export * from './readme';
 export * from './reserved-words';
 export * from './util';
+export * from './file-writer';
