+++ packages/openapi-generator/src/file-writer/index.ts
@@ -1,2 +1,1 @@
-export * from './copy-file';
 export * from './create-file';
